+++ gamemodes/terrortown/gamemode/client/cl_main.lua
@@ -20,7 +20,7 @@ ttt_include("cl_fonts")
 
 ttt_include("sh_init")
 
-ttt_include("sh_replicated_cvar")
+ttt_include("sh_cvar_handler")
 
 ttt_include("sh_network_sync")
 ttt_include("sh_sprint")
@@ -171,6 +171,8 @@ function GM:InitPostEntity()
 		CreateEquipment(itm) -- init items
 
 		itm.CanBuy = {} -- reset normal items equipment
+
+		itm:Initialize()
 	end
 
 	local sweps = weapons.GetList()