+++ gamemodes/terrortown/gamemode/server/sv_main.lua
@@ -6,7 +6,7 @@ include("ttt2/libraries/entity_outputs.lua")
 
 ttt_include("sh_init")
 
-ttt_include("sh_replicated_cvar")
+ttt_include("sh_cvar_handler")
 
 ttt_include("sh_sprint")
 ttt_include("sh_main")
@@ -345,6 +345,8 @@ function GM:InitPostEntity()
 		CreateEquipment(itm) -- init items
 
 		itm.CanBuy = {} -- reset normal items equipment
+
+		itm:Initialize()
 	end
 
 	for i = 1, #sweps do